fix!(fock): Rewrite `PostSelectPhotons`
This is not really a fix, rather a rewrite:
- `PostSelectPhotons` can be considered as a subclass of `Measurement`, because
  generally we cannot use the same `Simulator` for further simulations.
- `PostSelectPhotons` should really be returning with the state on the
  remaining modes, it is not needed to keep the postselected photons in the
  ancilla mode.
