+++ docs/index.rst
@@ -73,7 +73,6 @@ To install Piquasso with JAX is done by
    instructions/gates
    instructions/measurements
    instructions/channels
-   instructions/misc
 
 .. toctree::
    :maxdepth: 3
