--- docs/instructions/misc.rst
@@ -1,5 +1,0 @@
-Miscellaneous
-=============
-
-.. automodule:: piquasso.instructions.misc
-   :members:
