+++ tests/backends/tensorflow/test_gates.py
@@ -16,7 +16,6 @@
 import piquasso as pq
 import pytest
 import tensorflow as tf
-import numpy as np
 
 
 @pytest.mark.monkey
@@ -36,73 +35,3 @@ def test_Interferometer_numpy_array_as_parameter(generate_unitary_matrix):
         pq.Q(all) | pq.Interferometer(interferometer)
 
     simulator.execute(program)
-
-
-def test_PostSelectPhotons_gradient():
-    def _calculate_loss(weights, calculator, state_vector):
-        np = calculator.np
-
-        with pq.Program() as preparation:
-            pq.Q(all) | pq.StateVector([0, 1, 0]) * state_vector[0]
-            pq.Q(all) | pq.StateVector([1, 1, 0]) * state_vector[1]
-            pq.Q(all) | pq.StateVector([2, 1, 0]) * state_vector[2]
-
-        phase_shifter_phis = weights[:3]
-        thetas = weights[3:6]
-        phis = weights[6:]
-        with pq.Program() as interferometer:
-            for i in range(3):
-                pq.Q(i) | pq.Phaseshifter(phase_shifter_phis[i])
-
-            pq.Q(1, 2) | pq.Beamsplitter(theta=thetas[0], phi=phis[0])
-            pq.Q(0, 1) | pq.Beamsplitter(theta=thetas[1], phi=phis[1])
-            pq.Q(1, 2) | pq.Beamsplitter(theta=thetas[2], phi=phis[2])
-
-        with pq.Program() as program:
-            pq.Q(all) | preparation
-
-            pq.Q(all) | interferometer
-
-            pq.Q(all) | pq.PostSelectPhotons(
-                postselect_modes=(1, 2), photon_counts=(1, 0)
-            )
-
-        simulator = pq.PureFockSimulator(
-            d=3, config=pq.Config(cutoff=4), calculator=calculator
-        )
-
-        state = simulator.execute(program).state
-
-        state.normalize()
-
-        reduced_state = state.reduced((0,))
-
-        density_matrix = reduced_state.density_matrix[:3, :3]
-
-        expected_state = np.copy(state_vector)
-        expected_state = calculator.assign(expected_state, 2, -expected_state[2])
-
-        loss = 1 - np.sqrt(
-            np.real(np.conj(expected_state) @ density_matrix @ expected_state)
-        )
-
-        return loss
-
-    calculator = pq.TensorflowCalculator()
-
-    weights = tf.Variable(
-        [np.pi, 0.0, 0.0, np.pi / 8, 65.5302 * 2 * np.pi / 360, -np.pi / 8, 0, 0, 0]
-    )
-
-    with tf.GradientTape() as tape:
-        loss = _calculate_loss(
-            weights=weights,
-            calculator=calculator,
-            state_vector=np.sqrt([0.2, 0.3, 0.5]),
-        )
-
-    grad = tape.gradient(loss, weights)
-
-    assert np.isclose(loss, 0.0)
-
-    assert np.allclose(grad, 0.0, atol=1e-7)
